Fix task 5

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
 */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int res = 0, num = 0, temp = 0;
+	unsigned long int res = 0, num = 0, temp = 0;
 
 	num = n ^ m;
 	for (; 1==1 ;)
